feat: improve hot loop logging
