+++ yarn-project/archiver/src/archiver/index.ts
@@ -1,6 +1,5 @@
 export * from './archiver.js';
 export * from './config.js';
-export * from './epoch_helpers.js';
 export { type L1Published, type L1PublishedData } from './structs/published.js';
 export { MemoryArchiverStore } from './memory_archiver_store/memory_archiver_store.js';
 export { ArchiverDataStore } from './archiver_store.js';
