+++ yarn-project/circuit-types/src/index.ts
@@ -24,3 +24,4 @@ export * from './simulation_error.js';
 export * from './tx/index.js';
 export * from './tx_effect.js';
 export * from './tx_execution_request.js';
+export * from './epoch_helpers/index.js';
